backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.124.2",
    "httpx>=0.28.1",
    "mysqlclient>=2.2.7",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "reportlab>=4.4.6",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
